refactor(get-started): extract shared useScrollToHash hook
Dedupe the identical hash scroll-and-highlight effect from the plan and
setup onboarding routes into a shared ~/hooks/useScrollToHash hook, so
future deep-linked onboarding routes reuse one implementation instead of
copying the effect.
